+++ packages/resume/changelog.md
@@ -1,5 +1,6 @@
 # Unreleased
 * Add combinators `resumeOr` and `resumingOr`, which take an additional branch for the success case.
+* Add operator versions of `resume` (`!!`) and `resumeAs` (`<!`).
 
 # 0.1.0.0
 * initial release